+++ src/main/java/net/rasanovum/rockandstone/util/OreScanner.java
@@ -12,7 +12,11 @@ import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.ChunkAccess;
-import net.minecraft.world.level.chunk.ChunkStatus;
+//? if <1.21 {
+/*import net.minecraft.world.level.chunk.ChunkStatus;
+*///?} else {
+import net.minecraft.world.level.chunk.status.ChunkStatus;
+//?}
 import net.minecraft.world.level.levelgen.DensityFunction;
 import net.minecraft.world.level.levelgen.NoiseRouter;
 import net.rasanovum.rockandstone.RockAndStoneConfig;